refactor(menu): extract term init/restore funcs to file
