+++ src/terminal_utils.rs
@@ -1,0 +1,33 @@
+//!
+//!
+use std::io;
+
+use crossterm::{
+    execute,
+    terminal::{
+        EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode,
+        enable_raw_mode,
+    },
+};
+use ratatui::{DefaultTerminal, Terminal, prelude::CrosstermBackend};
+
+use anyhow::Result;
+
+/// Initializes the terminal in raw mode and alternate screen.
+///
+/// Returns a [`DefaultTerminal`] that must later be passed to [`restore`].
+pub fn init() -> Result<DefaultTerminal> {
+    enable_raw_mode()?;
+    let mut stdout = io::stdout();
+    execute!(stdout, EnterAlternateScreen)?;
+    let backend = CrosstermBackend::new(stdout);
+    let terminal = Terminal::new(backend)?;
+    Ok(terminal)
+}
+
+/// Restores the terminal to its normal mode and leaves the alternate screen.
+pub fn restore(mut terminal: DefaultTerminal) -> Result<()> {
+    disable_raw_mode()?;
+    execute!(terminal.backend_mut(), LeaveAlternateScreen)?;
+    Ok(())
+}
